Put default back button behavior in base class
